IPMI FRU image tool
`ipmi-fru-it` is a tool that can be used for reading/writing IPMI FRU
images.

It takes a config file as input and generates a IPMI FRU binary file
containing data from the config file. This binary file can be written
onto a FRU EEPROM. The config file is in a simple INI format (no
sub-sections).

`ipmi-fru-it` follows the IPMI FRU specifications for generating the
FRU data. A sample config file `fru.conf.sample`, is provided for
reference.

Refer to README for a list of known issues.